+++ spacecraft-common/src/com/treehouse/spacecraft/core/data/entity/DefaultEntity.java
@@ -7,7 +7,7 @@ public class DefaultEntity extends AbstractEntity{
 	 */
 	private static final long serialVersionUID = 4270207197248094915L;
 
-	public DefaultEntity(double x, double y, double width, double height) {
+	public DefaultEntity(float x, float y, float width, float height) {
 		super(x, y, width, height);
 	}
